Update int double convertion

+++ examples/load-balance/conga-simulation-large.cc
@@ -659,7 +659,7 @@ int main (int argc, char *argv[])
         }
     }
 
-    double oversubRatio = (SERVER_COUNT * LEAF_SERVER_CAPACITY) / (SPINE_LEAF_CAPACITY * SPINE_COUNT * LINK_COUNT);
+    double oversubRatio = static_cast<double>(SERVER_COUNT * LEAF_SERVER_CAPACITY) / (SPINE_LEAF_CAPACITY * SPINE_COUNT * LINK_COUNT);
     NS_LOG_INFO ("Over-subscription ratio: " << oversubRatio);
 
     NS_LOG_INFO ("Initialize CDF table");
